herosection van video voorzien, herosection tekst geredigeerd, daarna styling opgezet om de video naar de achtergrond te verplaatsen, section input component aangepast
