+++ src/components/HeroVideo/HeroVideo.module.css
@@ -1,0 +1,9 @@
+.video__background{
+    position: absolute;
+    z-index: -1;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    min-width: 100%;
+    min-height: 100%;
+}
